% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pkgdown_favicons.R
\name{build_pkgdown_favicons}
\alias{build_pkgdown_favicons}
\title{A local alternative to \link[pkgdown:build_favicons]{pkgdown::build_favicons}}
\usage{
build_pkgdown_favicons(pkg = ".", overwrite = FALSE)
}
\arguments{
\item{pkg}{Path to package.}

\item{overwrite}{If \code{TRUE}, re-create favicons from package logo.}
}
\description{
The list of returned icons is exactly the same as
\code{\link[pkgdown:build_favicons]{pkgdown::build_favicons()}}, to allow this function to be used as a drop-in
replacement.
}
