% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgicon-package.R
\docType{package}
\name{pkgicon-package}
\alias{pkgicon}
\alias{pkgicon-package}
\title{pkgicon: A Local and Robust Way to Generate Favicons for Package Documentation Sites}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [copyright holder]

}
\keyword{internal}
